+++ backup/backup_summary.sh
@@ -3,40 +3,67 @@
 source ./utils.sh
 
 function backup() {
-    if [[ ! -d "$2" ]]; then
-        mkdirprint "$2";
-    fi
+    # Variables for Summary
+    local ERRORS="0"
+    local WARNINGS="0"
+    local FILES_UPDATED="0"
+    local FILES_COPIED="0"
+    local FILES_DELETED="0"
+    local SIZE_COPIED="0"
+    local SIZE_REMOVED="0"
     for file in "$1"/*; do
         if is_in_list "$file" "${DIRS[@]}" ; then
             continue;
         fi
         if [[ -d "$file" ]]; then
+            mkdirprint "$2/$(basename "$file")";
             backup "$file" "$2/$(basename "$file")"
             continue;
         elif [[ ! "$(basename "$file")" =~ $REGEX ]]; then
             continue;
         fi
-        cpprint_summary "$file" "$2/$(basename "$file")"
+        local file_copy="$2/$(basename "$file")"
+        local simpler_name_workdir="${file#$(dirname "$WorkDir")/}"
+        local simpler_name_backup="${file_copy#$(dirname "$Backup")/}"
+        local FILE_MODE_DATE=$(stat -c %Y "$file")
+        if [ -f "$file_copy" ]; then
+            local BAK_FILE_DATE=$(stat -c %Y "$file_copy")
+            if [[ "$FILE_MODE_DATE" -lt "$BAK_FILE_DATE" ]]; then
+                echo "WARNING: backup entry $simpler_name_backup is newer than $simpler_name_workdir; Should not happen"
+                ((WARNINGS++))
+                continue;
+            elif [[ "$FILE_MODE_DATE" -eq "$BAK_FILE_DATE" ]]; then
+                continue;
+            fi
+            ((FILES_UPDATED++))
+        fi
+        ((FILES_COPIED++))
+        ((SIZE_COPIED+=$(stat -c %s "$file")))
+        echo "cp -a "$simpler_name_workdir" "$simpler_name_backup""
+        if [[ $CHECKING -eq 0 ]]; then
+            cp -a "$file" "$file_copy";
+        fi
     done
-}
-
-function backup_delete() {
-    if [[ ! -d "$2" || ! -n "$2" ]]; then
+    if [[ ! -d "$2" ]]; then
+        summary "$1" "$ERRORS" "$WARNINGS" "$FILES_UPDATED" "$FILES_COPIED" "$SIZE_COPIED" "$FILES_DELETED" "$SIZE_REMOVED"
         return 0;
     fi
+
     for file in "$2"/*; do
         if is_in_list "$file" "${DIRS[@]}" ; then
             continue;
         fi
         if [[ -d "$file" ]]; then
-            if [[ ! -d "$1/$(basename "$file")" ]]; then
+            if [[ ! -d "$1/$(basename "$file")" && $CHECKING -eq "0" ]]; then
+                local directory_size=$(du -sk "$file" | awk '{print $1}')
+                ((SIZE_REMOVED+=$directory_size))
+                local file_count=$(find "$file" -type f | wc -l)
+                ((FILES_DELETED+=$file_count))
                 rm -rf "$file"
-                continue;
             fi
-            backup_delete "$1/$(basename "$file")" "$2/$(basename "$file")"
             continue;
         fi 
-        if [[ ! -f "$file" || -f "$1/$(basename "$file")" ]]; then
+        if [[ -f "$1/$(basename "$file")" ]]; then
             continue;
         fi
         ((SIZE_REMOVED+=$(stat -c %s "$file") ))
@@ -45,23 +72,17 @@ function backup_delete() {
             rm "$file"
         fi
     done
-}
 
-# Variables for Summary
-ERRORS="0"
-WARNINGS="0"
-FILES_UPDATED="0"
-FILES_COPIED="0"
-FILES_DELETED="0"
+    summary "$1" "$ERRORS" "$WARNINGS" "$FILES_UPDATED" "$FILES_COPIED" "$SIZE_COPIED" "$FILES_DELETED" "$SIZE_REMOVED"
+}
 
 # Variables for the opts
 CHECKING="0"
 DIRS_FILE=""
 REGEX=""
 DIRS=()
-SIZE_COPIED="0"
-SIZE_REMOVED="0"
 
+OPTERR=0 
 while getopts "cb:r:" opt; do
     case $opt in
         c)
@@ -85,49 +106,78 @@ while getopts "cb:r:" opt; do
             REGEX="$OPTARG"
             check_regex "$REGEX"
             if [[ $? -eq 1 ]]; then
-                ((ERRORS++))
-                summary
-                exit 1
+                REGEX="(a"
             fi
             ;;
         \?)
-            echo "Invalid option: -$OPTARG"
-            exit 1
+            echo "ERROR: Invalid option selected"
+            REGEX="(a"
             ;;
         :)
             echo "Option -$OPTARG requires an argument."
-            exit 1
+            REGEX="(a"
             ;;
     esac
 done
 
 shift $((OPTIND - 1))
-
-if [[ ! -d "$1" ]]; then
-    echo "$1 not a dir"
+if [[ $# -lt 2 ]]; then
+    echo "ERROR: Not enough arguments"
+    summary "$1" "1" "0" "0" "0" "0" "0" "0"
+    exit 1
+elif [[ "$REGEX" == "(a" ]]; then
+    summary "$1" "1" "0" "0" "0" "0" "0" "0"
+    exit 1
+elif [[ ! -d "$1" ]]; then
+    echo "ERROR: "$(basename "$1")" is not a directory"
+    summary "$1" "1" "0" "0" "0" "0" "0" "0"
     exit 1;
 fi
 
-if [[ ! -d "$2" ]]; then
-    mkdirprint "$2";
-fi
+mkdirprint "$2" "$2";
 
 
-WORKDIR="$(realpath "$1")"
-BACKUP="$(realpath "$2")"
-BACKUP_PATH="$BACKUP"
+WorkDir="$(realpath "$1")"
+Backup="$(realpath "$2")"
+BackupPath="$Backup"
 
-while [[ "$BACKUP_PATH" != "/" ]]; do
-    if [[ $WORKDIR == $BACKUP_PATH ]]; then
-        echo "ERROR: $WORKDIR is a parent to $BACKUP"
-        ((ERRORS++))
-        summary
+# Calculate the total size of files in the source directory (in KB)
+WorkDirSize=$(du -sk "$WorkDir" | awk '{print $1}')
+
+if [[ -d "$2" ]]; then
+
+    # Get available space in the destination directory (in KB)
+    AvailableSpace=$(df -k "$Backup" | awk 'NR==2 {print $4}')
+
+    # Check if there's enough space in the destination directory
+    if (( AvailableSpace < WorkDirSize )); then
+        echo "ERROR: Not enough space in destination directory."
+        summary "$(basename "$WorkDir")" "1" "0" "0" "0" "0" "0" "0"
         exit 1
     fi
-    BACKUP_PATH="$(dirname "$BACKUP_PATH")"
+else
+
+    # Get available space in the computer (in KB)
+    AvailableSpace=$(df -k "/" | awk 'NR==2 {print $4}')
+
+    # Check if there's enough space in the destination directory
+    if (( AvailableSpace < WorkDirSize )); then
+        echo "ERROR: Not enough space in the computer."
+        summary "$(basename "$WorkDir")" "1" "0" "0" "0" "0" "0" "0"
+        exit 1
+    fi
+fi
+
+while [[ "$BackupPath" != "/" ]]; do
+    if [[ $WorkDir == $BackupPath ]]; then
+        echo "ERROR: "$(basename "$WorkDir")" is parent of "$(basename "$Backup")""
+        summary "$(basename "$WorkDir")" "1" "0" "0" "0" "0" "0" "0"
+        exit 1
+    fi
+    BackupPath="$(dirname "$BackupPath")"
 done
 
-backup_delete "$WORKDIR" "$BACKUP"
-backup "$WORKDIR" "$BACKUP"
-
-summary
+shopt -s nullglob dotglob
+backup "$WorkDir" "$Backup"
+echo $file_count
+exit 0
